+++ docs/source/News.rst
@@ -38,6 +38,8 @@ Backwards-incompatible changes
   to the backwards incompatible change on 
   :func:`repoze.what.middleware.setup_auth` introduced in 
   :ref:`repoze.what-1.0b2`.
+* Now it's possible to customize the authentication/identification cookie
+  through :func:`repoze.what.plugins.quickstart.setup_sql_auth`.
 
 .. _repoze.what-1.0b1: